fix(ssq): cnn_math 训练 dtype 修复 + load_latest_probs 每模型取最新 run_at
1. ml/main.py: cnn_math 训练分支 fm 缺 .astype(np.float32), 302列混合 int/float
   导致 dfe[fc].values 返回 object array, torch.tensor(...,float32) 抛
   TypeError。补 .astype(np.float32) 与 predict 分支一致。
2. select_numbers.py: load_latest_probs 原取 '全局 MAX(run_at)' 单一 run,
   导致模型分开重训时集成退化为单模型。改为每模型每球种独立取 MAX(run_at),
   避免 '必须一次跑完所有模型共享 run_at' 的耦合。

验证: cnn_math 重训成功(31.7s 入PG); 6模型集成(等权+EBMA)均正确取到各自最新。

+++ ml/main.py
@@ -72,7 +72,7 @@ def run_train(mt, df, retrain=True, col=None):
         tc = RED_COLS + BLUE_COLS; rt = m.config.get("regression_targets", ["Next_Sum","Next_OddRatio","Next_BigRatio","Next_Hot_Count","Next_Cold_Count","Next_Max_Omission","Next_Avg_Omission"])
         if not all(c in dfe.columns for c in rt): dfe = fe.generate_regression_targets(dfe)
         yc = np.array([[r-(i+1) for i,r in enumerate(row[:6])] + [row[6]-1] for row in dfe[tc].values])
-        yr = dfe[rt].values; fm = dfe[extract_feature_columns(dfe)].values
+        yr = dfe[rt].values; fm = dfe[extract_feature_columns(dfe)].values.astype(np.float32)
         X, yc, yr = m.prepare_data(fm, yc, yr)
         X, _, yc, _, yr, _ = train_test_split(X, yc, yr, test_size=0.2, random_state=42)
